Refactored genericCore files
